Remove redundant base path in manifest
Let's say that gulp.src() was invoked like this:

gulp.src([
    'app/assets/js/**/*.js', // first source dir
    'app/assets/vendor/**/*.js' // second source dir
])

In this case, firstFile would be equal to a first file encountered, and
it would be a file from first directory. That would cause relPath() to
return absolute path to a resource for each file that comes from a
second or any other subsequent directory.

As a result, manifest json would look like:

{
    // comes from the first source dir
    'app.js': 'app-01020304.js',
    'test.js': 'test-01020304.js',

    // comes from second source dir
    '/home/foo/dev/app/assets/vendor/moxie.js': 'moxie-01020304.js',
    '/home/foo/dev/app/assets/vendor/bar/quux.js': 'bar/quux-01020304.js'
}

This patch changes that behavior. Instead of relying on first
encountered file's base path for all remaining files, manifest key is
determined by combining original file name, and a relative path name
used to save revved file into a directory given to gulp.dest().