Make tests independent of platform specific path separator (#219)
Most of the path module's method will replace forward and backward
slashes with the platform specific path separator. And gulp-rev enforces
the use of forward slashes for the manifest. The tests didn't take these
two behaviors into account, which resulted in mismatches for some test
on platforms that use backward slashes as the path separator (i.e. the
Windows platform).