Add options to forceRev
forceRev is useful when running in dev mode. You might want to generate
the manifest file but you might not want to alter it if you are running
a watch on some files, as this would generate a new index.html file as
the paths are updated, this can make browsersync reload both the html
and the css, breaking the style reloading.

This could apply to more situations where you want the manifest, but not
the hash part in certain situations.

Added a test for this option.